src/main/java/chess/repository/MemoryMemberRepository.java
package chess.repository;

import chess.Member;
import java.util.ArrayList;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import java.util.concurrent.ConcurrentHashMap;

public class MemoryMemberRepository implements MemberRepository {
    private static Map<Long, Member> store = new ConcurrentHashMap<>();
    private static int nextId = 1;

    @Override
    public void save(Member member) {
        member = new Member((long) nextId++, member.getName());
        store.put(member.getId(), member);
    }

    @Override
    public Optional<Member> findById(Long id) {
        return Optional.of(store.get(id));
    }

    @Override
    public List<Member> findAll() {
        return new ArrayList<>(store.values());
    }

    void deleteAll() {
        store.clear();
    }
}
